+++ resources/[dg]/dg-hospital/typescript/src/client/modules/health/controller.health.ts
@@ -10,3 +10,5 @@ Events.onNet('hospital:health:useHeal', (healthIncrease: number, bleedDecrease: 
   const currentBleed = getBleedAmount();
   setBleedAmount(currentBleed - bleedDecrease);
 });
+
+global.exports('setHealth', setHealth);